+++ examples/apollo-client/src/index.js
@@ -1,4 +1,4 @@
-import { enableTracing } from '@envy/web';
+import { enableTracing } from '@envyjs/web';
 import { createRoot } from 'react-dom/client';
 
 import { App } from './App';