+++ packages/webui/src/types/index.ts
@@ -1,4 +1,4 @@
-import { Event } from '@envy/core';
+import { Event } from '@envyjs/core';
 
 export type Trace = Event;
 export type Traces = Map<string, Trace>;