+++ packages/client/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @envy/client
+# @envyjs/client
 
 ## 0.1.1
 
@@ -7,4 +7,4 @@
 - ba5e2b9: Add changesets release management
 - Updated dependencies [111807a]
 - Updated dependencies [ba5e2b9]
-  - @envy/core@0.2.0
+  - @envyjs/core@0.2.0