+++ packages/webui/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @envy/webui
+# @envyjs/webui
 
 ## 0.1.1
 
@@ -7,5 +7,5 @@
 - ba5e2b9: Add changesets release management
 - Updated dependencies [111807a]
 - Updated dependencies [ba5e2b9]
-  - @envy/core@0.2.0
-  - @envy/client@0.1.1
+  - @envyjs/core@0.2.0
+  - @envyjs/client@0.1.1